# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Winston AI Companion API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ConversationMessageRole =
  (typeof ConversationMessageRole)[keyof typeof ConversationMessageRole];

export const ConversationMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ConversationMessage {
  role: ConversationMessageRole;
  content: string;
}

export interface ChatRequest {
  /** The user's message to Winston */
  message: string;
  /** Previous conversation messages */
  history?: ConversationMessage[];
}

export interface ChatResponse {
  /** Winston's text response */
  reply: string;
  /** Google Maps URL to open if the response includes navigation directions */
  navigationUrl?: string;
}

export interface SpeakRequest {
  /** The text to convert to speech */
  text: string;
}

export interface SpeakResponse {
  /** Base64-encoded MP3 audio data */
  audioBase64: string;
  /** MIME type of the audio (audio/mpeg) */
  mimeType: string;
}

export interface ErrorResponse {
  error: string;
}

export interface GoalStep {
  id: number;
  goal_id: number;
  step_text: string;
  order: number;
  completed: boolean;
  completed_at?: string | null;
}

export interface Goal {
  id: number;
  user_name: string;
  title: string;
  description?: string | null;
  created_at: string;
  completed_at?: string | null;
  steps: GoalStep[];
}

export interface CreateGoalRequest {
  title: string;
  description?: string;
}

export interface AddGoalStepRequest {
  step_text: string;
  order?: number;
}

export interface UpdateGoalStepRequest {
  completed: boolean;
}

export interface BreakdownGoalRequest {
  goal: string;
  conversation_history?: ConversationMessage[];
}

export type BreakdownGoalResponseType =
  (typeof BreakdownGoalResponseType)[keyof typeof BreakdownGoalResponseType];

export const BreakdownGoalResponseType = {
  question: "question",
  steps: "steps",
} as const;

export interface BreakdownGoalResponse {
  type: BreakdownGoalResponseType;
  content: string;
  steps?: string[] | null;
}

export type GetGoals200 = {
  goals: Goal[];
};

export type DeleteGoal200 = {
  ok: boolean;
  deleted: boolean;
};