# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Winston AI Companion API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: chat
    description: Winston AI companion chat
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /chat:
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message to Winston
      description: Sends a user message to Claude AI and returns Winston's text response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Winston's response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /speak:
    post:
      operationId: textToSpeech
      tags: [chat]
      summary: Convert text to speech
      description: Converts a text string to audio using ElevenLabs TTS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeakRequest"
      responses:
        "200":
          description: Audio data as base64-encoded MP3
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /goals:
    get:
      operationId: getGoals
      tags: [goals]
      summary: Get all goals
      description: Returns all active goals for the user with their steps and completion status
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: object
                properties:
                  goals:
                    type: array
                    items:
                      $ref: "#/components/schemas/Goal"
                required:
                  - goals
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a new goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalRequest"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /goals/breakdown:
    post:
      operationId: breakdownGoal
      tags: [goals]
      summary: AI goal breakdown
      description: Uses Claude to ask a clarifying question or return 3-7 Stoic action steps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BreakdownGoalRequest"
      responses:
        "200":
          description: Question or structured steps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreakdownGoalResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /goals/{id}:
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal and all its steps
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goal deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  deleted:
                    type: boolean
                required: [ok, deleted]
        "404":
          description: Goal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /goals/{id}/steps:
    post:
      operationId: addGoalStep
      tags: [goals]
      summary: Add a step to a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddGoalStepRequest"
      responses:
        "201":
          description: Created step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalStep"
        "404":
          description: Goal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /goals/{id}/steps/{stepId}:
    patch:
      operationId: updateGoalStep
      tags: [goals]
      summary: Mark a step complete or incomplete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: stepId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoalStepRequest"
      responses:
        "200":
          description: Updated step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalStep"
        "404":
          description: Step not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ChatRequest:
      type: object
      properties:
        message:
          type: string
          description: The user's message to Winston
        history:
          type: array
          description: Previous conversation messages
          items:
            $ref: "#/components/schemas/ConversationMessage"
      required:
        - message
    ConversationMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required:
        - role
        - content
    ChatResponse:
      type: object
      properties:
        reply:
          type: string
          description: Winston's text response
        navigationUrl:
          type: string
          description: Google Maps URL to open if the response includes navigation directions
      required:
        - reply
    SpeakRequest:
      type: object
      properties:
        text:
          type: string
          description: The text to convert to speech
      required:
        - text
    SpeakResponse:
      type: object
      properties:
        audioBase64:
          type: string
          description: Base64-encoded MP3 audio data
        mimeType:
          type: string
          description: MIME type of the audio (audio/mpeg)
      required:
        - audioBase64
        - mimeType
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    GoalStep:
      type: object
      properties:
        id:
          type: integer
        goal_id:
          type: integer
        step_text:
          type: string
        order:
          type: integer
        completed:
          type: boolean
        completed_at:
          type: string
          nullable: true
      required:
        - id
        - goal_id
        - step_text
        - order
        - completed
    Goal:
      type: object
      properties:
        id:
          type: integer
        user_name:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        created_at:
          type: string
        completed_at:
          type: string
          nullable: true
        steps:
          type: array
          items:
            $ref: "#/components/schemas/GoalStep"
      required:
        - id
        - user_name
        - title
        - created_at
        - steps
    CreateGoalRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
      required:
        - title
    AddGoalStepRequest:
      type: object
      properties:
        step_text:
          type: string
        order:
          type: integer
      required:
        - step_text
    UpdateGoalStepRequest:
      type: object
      properties:
        completed:
          type: boolean
      required:
        - completed
    BreakdownGoalRequest:
      type: object
      properties:
        goal:
          type: string
        conversation_history:
          type: array
          items:
            $ref: "#/components/schemas/ConversationMessage"
      required:
        - goal
    BreakdownGoalResponse:
      type: object
      properties:
        type:
          type: string
          enum: [question, steps]
        content:
          type: string
        steps:
          type: array
          items:
            type: string
          nullable: true
      required:
        - type
        - content