# lib/api-zod/src/generated/types/breakdownGoalRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Winston AI Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationMessage } from "./conversationMessage";

export interface BreakdownGoalRequest {
  goal: string;
  conversation_history?: ConversationMessage[];
}