# lib/api-zod/src/generated/types/breakdownGoalResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Winston AI Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { BreakdownGoalResponseType } from "./breakdownGoalResponseType";

export interface BreakdownGoalResponse {
  type: BreakdownGoalResponseType;
  content: string;
  steps?: string[] | null;
}