# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Winston AI Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatResponse {
  /** Winston's text response */
  reply: string;
  /** Google Maps URL to open if the response includes navigation directions */
  navigationUrl?: string;
}