# lib/api-zod/src/generated/types/createGoalRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Winston AI Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGoalRequest {
  title: string;
  description?: string;
}