# lib/api-zod/src/generated/types/goal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Winston AI Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { GoalStep } from "./goalStep";

export interface Goal {
  id: number;
  user_name: string;
  title: string;
  description?: string | null;
  created_at: string;
  completed_at?: string | null;
  steps: GoalStep[];
}